--- nextstep-mvc/src/main/java/nextstep/mvc/tobe/core/ComponentScanner.java
@@ -1,5 +1,0 @@
-package nextstep.mvc.tobe.core;
-
-public interface ComponentScanner {
-    Object scan();
-}
